Improves the onboarding layout for uncommon screen sizes.
The Onboarding layout was too static to allow for good usage with

uncommon screen sizes or high zoom levels. The new layout improves the base

case but also allows for scrollbars for screen sizes and zoom levels, that

we didn't intend the usage of it for. This makes the Onboarding usable

on whatever device (except way too small screens like smartphones, which

isn't the requirement anyway).