Stops scrollbar from appearing unnecessarily and reserves space for it.
The scrollbar appeared when you clicked one of the setting tiles, since

the design element in the background caused the container to overflow. Setting

the overflow to hidden while the card is open, fixes this issue.

The relatively new and handy css property "scrollbar-gutter: stable" is also

added, that reserves the needed space for the scrollbar, meaning that there is

no layout shift when it appears and disappears.